application/requirements.txt
Django>=1.8,<2.0
mysqlclient
djangorestframework
coreapi

